Fix "modified twice in a single render" issue
Without this fix we get an error when "promptIsSelectable" is not
specified and the submit button is pressed. This is due to the fact
that ember-one-way-select is updating "promptIsSelectable" inside
"didReceiveAttrs" to false when it's blank.

This issue should maybe be reported to ember-one-way-components, but I
didn't manage to create a small reproduction :(