Remove bower
It didn't need to be a runtime dependency, and `ember-try` installs bower itself when needed, so it doesn't need to be in `package.json` at all. Getting this to work did require running `yarn upgrade` to get the latest semver-compatible version of all the dependencies -- I'm not sure which one was required to get the tests to run successfully, but it's a reasonable upgrade to do in any case.